pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tdata-session-exporter"
version = "0.1.0"
description = "Export Telegram Desktop tdata sessions to Telethon string sessions"
authors = [
  { name = "Romdevv", email = "romdevv@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.7"
dependencies = [
  "telethon>=1.0",
  "opentele>=1.15.0",
  "python-dotenv>=0.19.0"
]

readme = "README.md"

[tool.setuptools]
packages = ["tdata_session_exporter"]
